Separate panels for layer defaults and general caching options
