+++ src/web/gwc/src/main/java/org/geoserver/gwc/web/GWCSettingsPage.html
@@ -14,6 +14,9 @@
         <div wicket:id="gwcServicesPanel"></div>
       </li>
       <li>
+        <div wicket:id="layerDefaultsPanel"></div>
+      </li>
+      <li>
         <div wicket:id="cachingOptionsPanel"></div>
       </li>
     </ul>
