+++ src/web/gwc/src/main/resources/GeoServerApplication.properties
@@ -23,7 +23,6 @@ GWCSettingsPage.menuTitle                  = Caching Defaults
 GWCSettingsPage.manageGridSets             = Manage Gridsets
 
 GWCSettingsPage.gwcServices= GWC Services
-GWCSettingsPage.cachingOptions=Default Caching Options
 GWCSettingsPage.gwcProvidedServices = Provided Services
 GWCSettingsPage.lockProvider=Tile locking mechanism
 GWCSettingsPage.lockProvider.noLock=No locking (not recommended)
@@ -41,10 +40,12 @@ GWCSettingsPage.defaultCacheFormats=Default Tile Image Formats for:
 GWCSettingsPage.defaultCacheFormatsVector=Vector Layers
 GWCSettingsPage.defaultCacheFormatsRaster=Raster Layers
 GWCSettingsPage.defaultCacheFormatsOther=Layer Groups
+GWCSettingsPage.defaultCachedGridsets = Default Cached Gridsets
 GWC.ImageIOFileCachingThresholdUnsetWarning=The "Coverage Access" page "ImageIO Cache Memory Threshold" is unset, \
    and can cause a severe performance penalty when seeding tiles. A value of 1024K should suffice for the most common tile sizes. 
 
-CachingOptionsPanel.cachedGridsets = Default Cached Gridsets
+
+CachingOptionsPanel.cacheOptions = General Caching Options
 CachingOptionsPanel.nioLock = NIO based lock (interprocess locking)
 CachingOptionsPanel.memoryLock = In-process locking
 CachingOptionsPanel.noLock = Disable locking
