+++ src/web/gwc/src/main/resources/GeoServerApplication_de.properties
@@ -40,9 +40,9 @@ GWCSettingsPage.defaultCacheFormats=Standardbildformate für den Cache\:
 GWCSettingsPage.defaultCacheFormatsVector=Vektorlayer
 GWCSettingsPage.defaultCacheFormatsRaster=Rasterlayer
 GWCSettingsPage.defaultCacheFormatsOther=Gruppenlayer
+GWCSettingsPage.defaultCachedGridsets = Standard Cache Gridsets
 GWC.ImageIOFileCachingThresholdUnsetWarning=Die "Coverage Access"-Seite "ImageIo Cache Memory Threshold" ist nicht gesetzt und kann ernsthafte Performanceverluste beim Vorberechnen der Kartenteile verursachen. Ein Wert von 1024K sollte für die meisten üblichen Teil-Größen ausreichen.
 
-CachingOptionsPanel.cachedGridsets = Standard Cache Gridsets
 CachingOptionsPanel.nioLock = NIO basierte Sperre (Inter-Prozess Sperre)
 CachingOptionsPanel.memoryLock = In-Prozess Sperre
 CachingOptionsPanel.noLock = Sperre deaktivieren
