+++ src/web/gwc/src/main/resources/GeoServerApplication_fr.properties
@@ -43,7 +43,7 @@ GWCSettingsPage.defaultCacheFormatsRaster=Couches rasters
 GWCSettingsPage.defaultCacheFormatsOther=WMS/Groupes de couches
 GWC.ImageIOFileCachingThresholdUnsetWarning=Le paramètre "Limite mémoire du cache ImageIO" de la page "Coverage Access" n'est pas défini, ce qui peut fortement diminuer la performance de la génération des tuiles. Une valeur de 1024K devrait être suffisante pour la plupart des tailles de tuile.
 
-CachingOptionsPanel.cachedGridsets = Grilles de cache par défaut
+GWCSettingsPage.defaultCachedGridsets = Grilles de cache par défaut
 CachingOptionsPanel.nioLock = Verrouillage basé sur NIO (inter-processus)
 CachingOptionsPanel.memoryLock = Verrouillage interne au processus
 CachingOptionsPanel.noLock = Désactiver le verrouillage
