+++ src/web/gwc/src/main/resources/GeoServerApplication_ja.properties
@@ -42,7 +42,7 @@ GWCSettingsPage.defaultCacheFormatsRaster=\u30e9\u30b9\u30bf\u30ec\u30a4\u30e4
 GWCSettingsPage.defaultCacheFormatsOther=\u30ec\u30a4\u30e4\u30b0\u30eb\u30fc\u30d7
 GWC.ImageIOFileCachingThresholdUnsetWarning="\u30ab\u30d0\u30ec\u30c3\u30b8\u30a2\u30af\u30bb\u30b9"\u30da\u30fc\u30b8\u306e"\u753b\u50cfIO\u30ad\u30e3\u30c3\u30b7\u30e5\u30e1\u30e2\u30ea\u3057\u304d\u3044\u5024"\u304c\u8a2d\u7f6e\u3055\u308c\u3066\u3044\u307e\u305b\u3093\u3002\u305d\u306e\u305f\u3081\u3001\u30bf\u30a4\u30eb\u3092\u30b7\u30fc\u30c9\u3059\u308b\u969b\u306b\u30d1\u30d5\u30a9\u30fc\u30de\u30f3\u30b9\u304c\u4f4e\u4e0b\u3057\u307e\u3059\u3002\u6700\u3082\u4e00\u822c\u7684\u306a\u30bf\u30a4\u30eb\u30b5\u30a4\u30ba\u306e\u5024\u3068\u3057\u3066\u306f1024K\u3067\u5341\u5206\u3067\u3059\u3002
 
-CachingOptionsPanel.cachedGridsets = \u30ad\u30e3\u30c3\u30b7\u30e5\u3055\u308c\u305f\u898f\u5b9a\u30b0\u30ea\u30c3\u30c9\u30bb\u30c3\u30c8
+GWCSettingsPage.defaultCachedGridsets = \u30ad\u30e3\u30c3\u30b7\u30e5\u3055\u308c\u305f\u898f\u5b9a\u30b0\u30ea\u30c3\u30c9\u30bb\u30c3\u30c8
 CachingOptionsPanel.nioLock = NIO\u30ed\u30c3\u30af (\u5185\u90e8\u30d7\u30ed\u30bb\u30b9\u30ed\u30c3\u30af)
 CachingOptionsPanel.memoryLock = \u30d7\u30ed\u30bb\u30b9\u5185\u30ed\u30c3\u30af
 CachingOptionsPanel.noLock = \u30ed\u30c3\u30af\u7121\u52b9\u5316
