+++ src/web/gwc/src/main/resources/GeoServerApplication_ko.properties
@@ -43,7 +43,7 @@ GWCSettingsPage.defaultCacheFormatsRaster=\ub798\uc2a4\ud130 \ub808\uc774\uc5b4
 GWCSettingsPage.defaultCacheFormatsOther=\ub808\uc774\uc5b4 \uadf8\ub8f9
 GWC.ImageIOFileCachingThresholdUnsetWarning="\ucee4\ubc84\ub9ac\uc9c0 \uc561\uc138\uc2a4"\ud398\uc774\uc9c0\uc758 "ImageIO\uc5d0\uc11c \uce90\uc2dc \uba54\ubaa8\ub9ac \uc784\uacc4\uac12"\uc744 \uc124\uc815\ud558\uc9c0 \uc54a\uc558\uc73c\uba70, \ud0c0\uc77c \uc0dd\uc131\uc2dc \uc2ec\uac01\ud55c \uc131\ub2a5 \uc800\ud558\ub97c \uc77c\uc73c\ud0ac \uc218 \uc788\uc2b5\ub2c8\ub2e4. 1024K \uac12\uc740 \uac00\uc7a5 \uc77c\ubc18\uc801\uc778 \ud0c0\uc77c \ud06c\uae30\ub97c \ub9cc\uc871\ud574\uc57c \ud569\ub2c8\ub2e4.
 
-CachingOptionsPanel.cachedGridsets = \uae30\ubcf8 \uce90\uc2dc\ub41c \uadf8\ub9ac\ub4dc\uc14b
+GWCSettingsPage.defaultCachedGridsets = \uae30\ubcf8 \uce90\uc2dc\ub41c \uadf8\ub9ac\ub4dc\uc14b
 CachingOptionsPanel.nioLock = NIO \uae30\ubc18 \uc7a0\uae08 (interprocess \uc7a0\uae08)
 CachingOptionsPanel.memoryLock = In-process \uc7a0\uae08
 CachingOptionsPanel.noLock = \uc7a0\uae08 \ud574\uc81c
