+++ src/web/gwc/src/main/resources/GeoServerApplication_nl.properties
@@ -41,7 +41,7 @@ GWCSettingsPage.defaultCacheFormatsOther = Groepen
 GWC.ImageIOFileCachingThresholdUnsetWarning= De geheugendrempel voor ImageIO (zie: Rasterverwerking) is niet ingesteld. \
     Dit kan leiden tot geheugenproblemen bij het genereren van tegels. Een waarde van tenminste 1024 dient te worden ingevuld. 
 
-CachingOptionsPanel.cachedGridsets = Grids
+GWCSettingsPage.defaultCachedGridsets = Grids
 CachingOptionsPanel.nioLock = NIO (interproces)
 CachingOptionsPanel.memoryLock = Tijdens de verwerking
 CachingOptionsPanel.noLock = Niet actief
