+++ src/web/gwc/src/main/resources/GeoServerApplication_pl.properties
@@ -43,7 +43,7 @@ GWCSettingsPage.defaultCacheFormatsRaster=Warstwy Rastrowe
 GWCSettingsPage.defaultCacheFormatsOther=Grupa warstw
 GWC.ImageIOFileCachingThresholdUnsetWarning=The "Coverage Access" page "ImageIO Cache Memory Threshold" is unset, \nand can cause a severe performance penalty when seeding tiles. A value of 1024K should suffice for the most common tile sizes.
 
-CachingOptionsPanel.cachedGridsets = Domy\u015blne Buforowanie Siatek Kartograficznych
+GWCSettingsPage.defaultCachedGridsets = Domy\u015blne Buforowanie Siatek Kartograficznych
 CachingOptionsPanel.nioLock = Blokowanie typu NIO (blokowanie mi\u0119dzyprocesowe)
 CachingOptionsPanel.memoryLock = Blokowanie w procesie
 CachingOptionsPanel.noLock = Wy\u0142\u0105cz blokowanie
