+++ src/web/gwc/src/main/resources/GeoServerApplication_ro.properties
@@ -43,7 +43,7 @@ GWCSettingsPage.defaultCacheFormatsRaster=Straturi raster
 GWCSettingsPage.defaultCacheFormatsOther=Grupuri de straturi
 GWC.ImageIOFileCachingThresholdUnsetWarning=Op\u021biunea 'Prag Memorie de Stocare ImageIO" din pagina "Acces la acoperire" nu este setat\u0103, ceea ce poate duce la o penalizare de performan\u021b\u0103 sever\u0103 atunci când se genereaz\u0103 dale. O valoare de 1024K ar trebui s\u0103 fie suficient\u0103 pentru dimensiunile comune ale unei dale.
 
-CachingOptionsPanel.cachedGridsets = Seturile implicite de grile înmagazinate
+GWCSettingsPage.defaultCachedGridsets = Seturile implicite de grile înmagazinate
 CachingOptionsPanel.nioLock = Blocare bazat\u0103 pe NIO (blocare interprocesual\u0103)
 CachingOptionsPanel.memoryLock = Blocare în-proces
 CachingOptionsPanel.noLock = Dezactiveaz\u0103 blocare
