+++ src/web/gwc/src/main/resources/GeoServerApplication_ru.properties
@@ -40,7 +40,7 @@ GWCSettingsPage.defaultCacheFormatsRaster=\u0420\u0430\u0441\u0442\u0440\u043e\u
 GWCSettingsPage.defaultCacheFormatsOther=WMS/\u0433\u0440\u0443\u043f\u043f\u044b \u0441\u043b\u043e\u0435\u0432
 GWC.ImageIOFileCachingThresholdUnsetWarning="\u041a\u043e\u043b\u0438\u0447\u0435\u0441\u0442\u0432\u043e \u043f\u0430\u043c\u044f\u0442\u0438 \u043a\u044d\u0448\u0430 ImageIO" \u043d\u0430 \u0441\u0442\u0440\u0430\u043d\u0438\u0446\u0435 "\u0414\u043e\u0441\u0442\u0443\u043f \u043a \u043f\u043e\u043a\u0440\u044b\u0442\u0438\u044f\u043c" \u043d\u0435 \u0443\u0441\u0442\u0430\u043d\u043e\u0432\u043b\u0435\u043d\u043e \u0438 \u043c\u043e\u0436\u0435\u0442 \u0432\u043b\u0435\u0447\u044c \u043a \u0441\u0435\u0440\u044c\u0435\u0437\u043d\u044b\u043c \u043f\u0440\u043e\u0431\u043b\u0435\u043c\u0430\u043c \u0441 \u043f\u0440\u043e\u0438\u0437\u0432\u043e\u0434\u0438\u0442\u0435\u043b\u044c\u043d\u043e\u0441\u0442\u044c\u044e \u043f\u0440\u0438 \u0432\u044b\u0431\u043e\u0440\u0435 \u0442\u0430\u0439\u043b\u043e\u0432. \u0417\u043d\u0430\u0447\u0435\u043d\u0438\u044f \u0432 1024\u041a \u043e\u0431\u044b\u0447\u043d\u043e \u0434\u043e\u0441\u0442\u0430\u0442\u043e\u0447\u043d\u043e \u0434\u043b\u044f \u0441\u0430\u043c\u044b\u0445 \u0440\u0430\u0441\u043f\u0440\u043e\u0441\u0442\u0440\u0430\u043d\u0435\u043d\u043d\u044b\u0445 \u0440\u0430\u0437\u043c\u0435\u0440\u043e\u0432 \u0442\u0430\u0439\u043b\u043e\u0432.
 
-CachingOptionsPanel.cachedGridsets = \u041a\u044d\u0448\u0438\u0440\u043e\u0432\u0430\u043d\u043d\u044b\u0435 \u0413\u0435\u043e\u0441\u0435\u0442\u043a\u0438 \u043f\u043e \u0443\u043c\u043e\u043b\u0447\u0430\u043d\u0438\u044e
+GWCSettingsPage.defaultCachedGridsets = \u041a\u044d\u0448\u0438\u0440\u043e\u0432\u0430\u043d\u043d\u044b\u0435 \u0413\u0435\u043e\u0441\u0435\u0442\u043a\u0438 \u043f\u043e \u0443\u043c\u043e\u043b\u0447\u0430\u043d\u0438\u044e
 CachingOptionsPanel.nioLock = \u0411\u043b\u043e\u043a\u0438\u0440\u043e\u0432\u043a\u0430 \u043d\u0430 \u043e\u0441\u043d\u043e\u0432\u0435 NIO (\u043c\u0435\u0436\u043f\u0440\u043e\u0446\u0435\u0441\u0441\u043d\u0430\u044f \u0431\u043b\u043e\u043a\u0438\u0440\u043e\u0432\u043a\u0430)
 CachingOptionsPanel.memoryLock = \u0412\u043d\u0443\u0442\u0440\u0438\u043f\u0440\u043e\u0446\u0435\u0441\u0441\u043d\u0430\u044f \u0431\u043b\u043e\u043a\u0438\u0440\u043e\u0432\u043a\u0430
 CachingOptionsPanel.noLock = \u041e\u0442\u043a\u043b\u044e\u0447\u0438\u0442\u044c \u0431\u043b\u043e\u043a\u0438\u0440\u043e\u0432\u043a\u0443
