+++ src/web/gwc/src/main/resources/GeoServerApplication_tr.properties
@@ -43,7 +43,7 @@ GWCSettingsPage.defaultCacheFormatsRaster=Raster Katmanlar
 GWCSettingsPage.defaultCacheFormatsOther=Katman Gruplar\u0131
 GWC.ImageIOFileCachingThresholdUnsetWarning="Coverage Eri\u015fim" sayfas\u0131 "ImageIO Ön Bellek E\u015fi\u011fi" kald\u0131r\u0131l\u0131rsa, karelerin yüklenmesinde ciddi performans kay\u0131plar\u0131 olabilir. 1024K de\u011feri en yayg\u0131n kare boyutlar\u0131 için yeterli olacakt\u0131r.
 
-CachingOptionsPanel.cachedGridsets = Varsay\u0131lan Önbelle\u011fe al\u0131nm\u0131\u015f Grid Dizileri
+GWCSettingsPage.defaultCachedGridsets = Varsay\u0131lan Önbelle\u011fe al\u0131nm\u0131\u015f Grid Dizileri
 CachingOptionsPanel.nioLock = NIO tabanl\u0131 bir kilit (i\u015flem aras\u0131 kilitleme)
 CachingOptionsPanel.memoryLock = \u0130\u015flem s\u0131ras\u0131nda kilitleme
 CachingOptionsPanel.noLock = Kilitlemeyi devre d\u0131\u015f\u0131 b\u0131rak
